Rename computed_props folder to computedProps
