+++ app/packs/src/components/computedProps/ComputedPropLabel.js
