+++ app/packs/src/components/computedProps/Highlight.js
