+++ app/packs/src/components/computedProps/NoComputedProps.js
