+++ app/packs/src/components/computedProps/SampleComputedProps.js
