+++ app/packs/src/components/elements/samples/SampleDetails.js
@@ -48,8 +48,8 @@ import ToggleSection from 'src/components/common/ToggleSection';
 import SampleName from 'src/components/common/SampleName';
 import ClipboardCopyText from 'src/components/common/ClipboardCopyText';
 import SampleForm from 'src/components/elements/samples/properties_tab/SampleForm';
-import ComputedPropsContainer from 'src/components/computed_props/ComputedPropsContainer';
-import ComputedPropLabel from 'src/components/computed_props/ComputedPropLabel';
+import ComputedPropsContainer from 'src/components/computedProps/ComputedPropsContainer';
+import ComputedPropLabel from 'src/components/computedProps/ComputedPropLabel';
 import Utils from 'src/utilities/Functions';
 import PrintCodeButton from 'src/components/common/PrintCodeButton';
 import SampleDetailsLiteratures from 'src/components/elements/DetailsTabLiteratures';
